Binding resolution refactored (#5802)
* Moved "node" and "react-native" into "src/platform"

* Renaming platform specific outputs from bindgen

* Using ts paths to resolve realm/binding

* Passing formatters as functions

* Ditch the typescript checker

* Simplified formatter naming

* Moved eslint formatter

Pulled in latest bindgen branch from core

* Temporarily update core

* Upgrading core to pull in the lint
