+++ packages/realm/.mocharc.json
@@ -2,6 +2,6 @@
   "loader": "tsx",
   "enable-source-maps": true,
   "no-warnings": true,
-  "require": "src/node/index.ts",
+  "require": "src/platform/node/index.ts",
   "spec": "src/tests/*.test.ts"
 }
