+++ packages/realm/bindgen/src/eslint-formatter.ts
@@ -1,0 +1,3 @@
+import { createCommandFormatter } from "@realm/bindgen/formatter";
+
+export const eslint = createCommandFormatter("eslint", ["npx", "eslint", "--fix", "--format=stylish"]);
