+++ packages/realm/bindgen/src/templates/jsi.ts
@@ -33,6 +33,7 @@ import {
 } from "@realm/bindgen/bound-model";
 
 import { doJsPasses } from "../js-passes";
+import { clangFormat } from "@realm/bindgen/formatter";
 
 // Code assumes this is a unique name that is always in scope to refer to the jsi::Runtime.
 // Callbacks need to ensure this is in scope. Functions taking Runtime arguments must use this name.
@@ -1040,7 +1041,7 @@ class JsiCppDecls extends CppDecls {
 }
 
 export function generate({ spec, file: makeFile }: TemplateContext): void {
-  const out = makeFile("jsi_init.cpp", "clang-format");
+  const out = makeFile("jsi_init.cpp", clangFormat);
 
   // HEADER
   out(`// This file is generated: Update the spec instead of editing this file directly`);
