--- packages/realm/src/node/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "types": ["node"],
-    "noResolve": false,
-    "incremental": true,
-    "resolveJsonModule": true
-  },
-  "exclude": [
-    "../tests/",
-    "../react-native/",
-    "../../type-tests/"
-  ],
-}
