+++ packages/realm/src/platform/react-native/index.ts
@@ -15,10 +15,9 @@
 // limitations under the License.
 //
 ////////////////////////////////////////////////////////////////////////////
-
 import "./fs";
 import "./device-info";
 
-export * from "../index";
-import { Realm } from "../index";
+export * from "../../index";
+import { Realm } from "../../index";
 export default Realm;
