--- packages/realm/src/react-native/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "noResolve": false,
-    "incremental": true,
-    "resolveJsonModule": true,
-    "types": [
-      "react-native",
-      "path-browserify"
-    ]
-  },
-  "exclude": [
-    "../tests/",
-    "../node/",
-    "../../type-tests/"
-  ],
-}
